embed_service/requirements.txt
protobuf==3.20.3
fastapi
uvicorn[standard]
sentence-transformers
numpy
python-dotenv
